# docs/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stoneboyz CRM API
  version: 0.1.0
  description: |
    Source of truth for all API contracts.
    Edit this file before implementing any endpoint.
    packages/api-client is generated from this file; do not hand-edit it.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.stoneboyz.app/v1
    description: Production placeholder
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
    CustomerId:
      name: customerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PriceListId:
      name: priceListId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PriceListItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        details:
          type: object
...
    PortalQuote:
      type: object
      required: [id, quoteNumber, title, status, customerId, customerName, discountCents, taxCents, subtotalCents, taxRateBps, shareToken, areas, lineItems]
      properties:
        id:
          type: string
          format: uuid
        quoteNumber:
          type: string
        title:
          type: string
        status:
          type: string
        customerId:
          type: string
          format: uuid
        customerName:
          type: string
        validUntil:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        termsAndConditions:
          type: string
          nullable: true
        discountCents:
          type: integer
        taxCents:
          type: integer
        subtotalCents:
          type: integer
        taxRateBps:
          type: integer
        shareToken:
          type: string
          format: uuid
        areas:
          type: array
          items:
            type: object
            required: [id, name, sortOrder]
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              sortOrder:
                type: integer
        lineItems:
          type: array
          items:
            type: object
            required: [id, stoneType, qty, qtyUnit, unitPriceCents, laborPriceCents, sortOrder]
            properties:
              id:
                type: string
                format: uuid
              stoneType:
                type: string
              qty:
                type: number
              qtyUnit:
                type: string
              unitPriceCents:
                type: integer
              laborPriceCents:
                type: integer
              lengthIn:
                type: number
                nullable: true
              widthIn:
                type: number
                nullable: true
              thicknessCm:
                type: number
                nullable: true
              edgeProfile:
                type: string
                nullable: true
              notes:
                type: string
                nullable: true
              sortOrder:
                type: integer
              quoteAreaId:
                type: string
                format: uuid
                nullable: true
    PriceListStatus:
      type: string
      enum: [draft, active, archived]
    PriceList:
      type: object
      required:
        - id
        - name
        - status
        - revision
        - currencyCode
        - defaultTaxRateBps
        - createdByUserId
        - archivedAt
        - archivedByUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PriceListStatus'
        revision:
          type: integer
        currencyCode:
          type: string
        defaultTaxRateBps:
          type: integer
        defaultPaymentTerms:
          type: string
          nullable: true
        expirationDays:
          type: integer
          nullable: true
        createdByUserId:
          type: string
          format: uuid
        archivedAt:
          type: string
          format: date-time
          nullable: true
        archivedByUserId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PriceListItem:
      type: object
      required:
        - id
        - priceListId
        - category
        - itemType
        - name
        - unit
        - priceCents
        - sortOrder
        - taxable
        - allowDiscount
        - editableOnQuote
        - hideOnQuote
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        priceListId:
          type: string
          format: uuid
        category:
          type: string
        itemType:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        unit:
          type: string
        priceCents:
          type: integer
        sortOrder:
          type: integer
        taxable:
          type: boolean
        allowDiscount:
          type: boolean
        editableOnQuote:
          type: boolean
        hideOnQuote:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PriceListWithItems:
      allOf:
        - $ref: '#/components/schemas/PriceList'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/PriceListItem'
...
    PriceListItemsResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PriceListItem'
...
    UpdatePriceListRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        revision:
          type: integer
        currencyCode:
          type: string
        defaultTaxRateBps:
          type: integer
        defaultPaymentTerms:
          type: string
          nullable: true
        expirationDays:
          type: integer
          nullable: true
    CreatePriceListItemRequest:
      type: object
      required: [category, itemType, name, unit, priceCents]
      properties:
        category:
          type: string
        itemType:
          type: string
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        priceCents:
          type: integer
        sortOrder:
          type: integer
        taxable:
          type: boolean
        allowDiscount:
          type: boolean
        editableOnQuote:
          type: boolean
        hideOnQuote:
          type: boolean
    UpdatePriceListItemRequest:
      type: object
      properties:
        category:
          type: string
        itemType:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        unit:
          type: string
        priceCents:
          type: integer
        sortOrder:
          type: integer
        taxable:
          type: boolean
        allowDiscount:
          type: boolean
        editableOnQuote:
          type: boolean
        hideOnQuote:
          type: boolean
    CustomerKind:
      type: string
      enum: [company, person]
    CustomerStatus:
      type: string
      enum: [lead, qualified, active, inactive, churned]
    CustomerType:
      type: string
      enum: [prospect, customer, partner, vendor]
...
    Customer:
      type: object
      required:
        - id
        - customerKind
        - name
        - status
        - type
        - ownerUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        customerKind:
          $ref: '#/components/schemas/CustomerKind'
        name:
          type: string
        companyName:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/CustomerStatus'
        type:
          $ref: '#/components/schemas/CustomerType'
        ownerUserId:
          type: string
          format: uuid
        primaryContactId:
          type: string
          format: uuid
          nullable: true
        billingContactId:
          type: string
          format: uuid
          nullable: true
        billingAddressId:
          type: string
          format: uuid
          nullable: true
        taxId:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        industry:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        notesSummary:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        whatsappPhone:
          type: string
          nullable: true
        billingEmail:
          type: string
          format: email
          nullable: true
        archiveReason:
          type: string
          nullable: true
        archivedAt:
          type: string
          format: date-time
          nullable: true
        archivedByUserId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateCustomerRequest:
      type: object
      minProperties: 1
      properties:
        customerKind:
          $ref: '#/components/schemas/CustomerKind'
        name:
          type: string
          minLength: 1
        companyName:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/CustomerStatus'
        type:
          $ref: '#/components/schemas/CustomerType'
        ownerUserId:
          type: string
          format: uuid
        taxId:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        industry:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        notesSummary:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        whatsappPhone:
          type: string
          nullable: true
        billingEmail:
          type: string
          format: email
          nullable: true
...
    CustomerAddressType:
      type: string
      enum: [billing, shipping, office, other]
    CustomerAddress:
      type: object
      required:
        - id
        - customerId
        - type
        - line1
        - city
        - country
        - isPrimary
        - isBilling
        - createdAt
        - updatedAt
        - archivedAt
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/CustomerAddressType'
        line1:
          type: string
        line2:
          type: string
          nullable: true
        city:
          type: string
        region:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
        isPrimary:
          type: boolean
        isBilling:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time
          nullable: true
    CreateCustomerAddressRequest:
      type: object
      required: [type, line1, city, country]
      properties:
        type:
          $ref: '#/components/schemas/CustomerAddressType'
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        region:
          type: string
        postalCode:
          type: string
        country:
          type: string
        isPrimary:
          type: boolean
          default: false
        isBilling:
          type: boolean
          default: false
    UpdateCustomerAddressRequest:
      type: object
      description: At least one updatable address field is required.
      minProperties: 2
      properties:
        type:
          $ref: '#/components/schemas/CustomerAddressType'
        line1:
          type: string
        line2:
          type: string
          nullable: true
        city:
          type: string
        region:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
        isPrimary:
          type: boolean
        isBilling:
          type: boolean
...
    ActorRequest:
      type: object
      properties:
...
security:
  - bearerAuth: []
paths:
...
  /price-lists/{priceListId}:
    get:
      operationId: getPriceList
      summary: Get a price list
      parameters:
        - $ref: '#/components/parameters/PriceListId'
      responses:
        '200':
          description: Price list returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceListWithItems'
    patch:
      operationId: updatePriceList
      summary: Update a draft price list
      parameters:
        - $ref: '#/components/parameters/PriceListId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePriceListRequest'
      responses:
        '200':
          description: Price list updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceList'
...
  /price-lists/{priceListId}/items:
    get:
      operationId: listPriceListItems
      summary: List price list items
      parameters:
        - $ref: '#/components/parameters/PriceListId'
      responses:
        '200':
          description: Items returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceListItemsResponse'
    post:
      operationId: createPriceListItem
      summary: Create a price list item
      parameters:
        - $ref: '#/components/parameters/PriceListId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePriceListItemRequest'
      responses:
        '201':
          description: Item created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceListItem'
  /price-lists/{priceListId}/items/{itemId}:
    patch:
      operationId: updatePriceListItem
      summary: Update a price list item
      parameters:
        - $ref: '#/components/parameters/PriceListId'
        - $ref: '#/components/parameters/PriceListItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePriceListItemRequest'
      responses:
        '200':
          description: Item updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceListItem'
    delete:
      operationId: deletePriceListItem
      summary: Hard-delete a price list item
      parameters:
        - $ref: '#/components/parameters/PriceListId'
        - $ref: '#/components/parameters/PriceListItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequest'
      responses:
        '200':
          description: Item deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceListItem'
...
  /customers/{customerId}:
    get:
      operationId: getCustomer
      summary: Get customer detail
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      operationId: updateCustomer
      summary: Update customer fields
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Customer conflicts with an existing non-archived customer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /customers/{customerId}/addresses:
    get:
      operationId: listCustomerAddresses
      summary: List customer addresses
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer addresses returned.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomerAddress'
        '404':
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      operationId: createCustomerAddress
      summary: Create customer address
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerAddressRequest'
      responses:
        '201':
          description: Customer address created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAddress'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Address conflicts with an existing active address rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/addresses/{addressId}:
    patch:
      operationId: updateCustomerAddress
      summary: Update customer address
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - name: addressId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerAddressRequest'
      responses:
        '200':
          description: Customer address updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAddress'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer address not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Address conflicts with an existing active address rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: archiveCustomerAddress
      summary: Archive customer address
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - name: addressId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequest'
      responses:
        '200':
          description: Customer address archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAddress'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer address not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/addresses/{addressId}/make-billing:
    post:
      operationId: makeAddressBilling
      summary: Make customer address billing
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - name: addressId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequest'
      responses:
        '200':
          description: Customer address made billing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAddress'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer address not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /portal/quotes/{token}:
    get:
      operationId: getPortalQuote
      summary: Get quote by share token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quote for customer portal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalQuote'
        '404':
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /portal/quotes/{token}/accept:
    post:
      operationId: acceptPortalQuote
      summary: Accept quote via share link (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /portal/quotes/{token}/reject:
    post:
      operationId: rejectPortalQuote
      summary: Reject quote via share link (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean